Relocate flag package
